src/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
flask-pony==3.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
kanpai==0.1.15
MarkupSafe==1.1.1
marshmallow==3.9.1
pony==0.7.13
pony-database-facade==2.0.1
pycodestyle==2.6.0
PyJWT==1.7.1
PyMySQL==0.10.1
python-decouple==3.3
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
Werkzeug==1.0.1